examples/gccflags: change flags to Enum to Boolean parameters
Choosing on, off or default for each flag adds some redundancy and
ambiguity to the search. This commit changes flags to BooleanParameters,
so that they can only be on (True) or off (False). The enabledInConfig
method can be used to avoid explicitly setting a flag that is already
the default.

This commit also sets the self._manipulator variable to avoid creating
a new manipulator everytime GccFlagsTuner.manipulator() is called.